requirements-pi.txt
blobfile==2.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
deepspeed==0.4.0
einops==0.6.0
exceptiongroup==1.1.1
fairscale==0.4.0
filelock==3.11.0
fonttools==4.39.3
ftfy==6.1.1
huggingface-hub==0.13.4
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
mypy==1.2.0
mypy-extensions==1.0.0
networkx==3.1
ninja==1.11.1
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
protobuf==3.20.3
psutil==5.9.4
pyarrow==11.0.0
pycocoevalcap==1.2
pycocotools==2.0.6
pycryptodomex==3.17
pyDeprecate==0.3.2
pyparsing==3.0.9
pytest==7.3.1
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
scipy==1.10.1
sentencepiece==0.1.98
six==1.16.0
sympy==1.11.1
tensorboardX==1.8
timm==0.4.12
tokenizers==0.13.3
tomli==2.0.1
torch @ https://download.pytorch.org/whl/torch-2.0.0-cp310-cp310-manylinux2014_aarch64.whl
torchmetrics==0.7.3
torchscale==0.2.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6